Trait pallet_author_inherent::pallet::Config

pub trait Config: Config {
    type AccountLookup: AccountLookup<Self::AccountId>;
    type CanAuthor: CanAuthor<Self::AccountId>;
    type SlotBeacon: SlotBeacon;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type AccountLookup: AccountLookup<Self::AccountId>

A type to convert between AuthorId and AccountId. This is useful when you want to associate Block authoring behavior with an AccoutId for rewards or slashing. If you do not need to hold an AccountID responsible for authoring use () which acts as an identity mapping.

type CanAuthor: CanAuthor<Self::AccountId>

The final word on whether the reported author can author at this height. This will be used when executing the inherent. This check is often stricter than the Preliminary check, because it can use more data. If the pallet that implements this trait depends on an inherent, that inherent must be included before this one.

type SlotBeacon: SlotBeacon

Some way of determining the current slot for purposes of verifying the author’s eligibility

type WeightInfo: WeightInfo